% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_parameters.R
\name{accumulate_parameters}
\alias{accumulate_parameters}
\title{Accumulate material properties into a prescribed list format}
\usage{
accumulate_parameters(name, sname, ac, acr, Mm)
}
\arguments{
\item{name}{character string; complete name of the material}

\item{sname}{character string; shortened name (e.g. atomic symbol, chemical name, etc) of the material}

\item{ac}{numeric vector; Antoine coefficents for vapor pressure calculations}

\item{acr}{numeric vector; temperature range \link{K} over which the Antoine coefficents are valid}

\item{Mm}{numeric vector; molar mass \link{kg/mol} of the material}
}
\description{
Accumulate material properties into a prescribed list format
}
\author{
Peter Norgard
}
