% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_material.R
\name{append_material}
\alias{append_material}
\title{Append Material}
\usage{
append_material(master, element)
}
\arguments{
\item{main_list}{list of lists; the main list of lists onto which the sub-list is to be appended}

\item{sub_list}{list; a sub-list containing one-for-one data under equivalent names}
}
\description{
Append material properties to the end of a list-of-lists containing specific properties of materials to be evaluated.
}
\author{
Peter Norgard
}
