% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aragonitekinetics.R
\name{aragonitekinetics}
\alias{aragonitekinetics}
\title{Aragonite Kinetics Interpolation Function}
\usage{
aragonitekinetics(tempC, reference = "Romanek")
}
\arguments{
\item{tempC}{temperature in degrees Celsius}

\item{reference}{string with the source of kinetics data - must be either "BurtonWalter" or "Romanek"}
}
\value{
a list with two items, k (umol/m\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}/hr) and n (unitless)
}
\description{
Calculates the precipitation rate parameters (k and n) for aragonite as a function of temperature.
}
\details{
This function interpolates between the experimental constrains of the rate constant (k) and reaction order (n) for aragonite precipitation as observed at 5,25, and 37 C in Burton and Walter (1987):

Burton, E. A., & Walter, L. M. (1987). Relative precipitation rates of aragonite and Mg calcite from seawater: Temperature or carbonate ion control? Geology, 15, 111–114.
or as observed at 10, 25, and 40 C in Romanek et al. (2011):
Romanek, C. S., Morse, J. W., & Grossman, E. L. (2011). Aragonite Kinetics in Dilute Solutions. Aquatic Geochemistry, 17(4), 339.
}
\examples{
arag1 <- aragonitekinetics(25)
arag2 <- mapply(aragonitekinetics,c(25,26,27))

}
