% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcitekinetics.R
\name{calcitekinetics}
\alias{calcitekinetics}
\title{Calcite Kinetics Interpolation Function}
\usage{
calcitekinetics(tempC, reference = "BurtonWalter")
}
\arguments{
\item{tempC}{temperature in degrees Celsius}

\item{reference}{string with the reference used for the calculation; the two options are Burton and Walter, 1989 ("BurtonWalter"), which is valid for 5 to 37 C, or Lopez et al. (2009) ("Lopezetal"), which is valid for 5 to 55 C.}
}
\value{
a list with two items, k (umol/m\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}/hr) and n (unitless)
}
\description{
Calculates the precipitation rate parameters (k and n) for calcite as a function of temperature.
}
\details{
This function interpolates between the experimental constrains of the rate constant (k) and reaction order (n) for calcite precipitation, with the choice of two references to use:

Burton, E. A., & Walter, L. M. (1987). Relative precipitation rates of aragonite and Mg calcite from seawater: Temperature or carbonate ion control? Geology, 15, 111–114.

Lopez, O., Zuddas, P., & Faivre, D. (2009). The influence of temperature and seawater composition on calcite crystal growth mechanisms and kinetics: Implications for Mg incorporation in calcite lattice. Geochimica et Cosmochimica Acta, 73(2), 337–347.
}
\examples{
calc1 <- calcitekinetics(25)
calc2 <- mapply(calcitekinetics,c(25,26,27))
calc3 <- calcitekinetics(25,datasource = "Lopezetal")

}
