% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matprop.R
\name{create_matprop}
\alias{create_matprop}
\title{Create material property list}
\usage{
create_matprop()
}
\description{
Generate a list of parameters that describes a material that will be subject to evaporation calculations.
}
\author{
Peter Norgard
}
