% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaporation.R
\name{evaporation}
\alias{evaporation}
\title{Evaporation Rate}
\usage{
evaporation(pv, p0, Tval, M, a = 1)
}
\arguments{
\item{pv}{numeric vector; vapor pressure \link{Pa} at which to evaluate the evaporation}

\item{p0}{numeric value; pressure \link{Pa} adjacent to evaporating surface}

\item{M}{numeric value; molar mass \link{kg/mol} of the evaporating material}

\item{a}{numeric value; sticking coefficient \link{unitless} with range \link{0,1} (default = 1)}

\item{T}{numeric vector; surface temperature \link{K} of the evaporating material}
}
\description{
Compute the evaporation rate of material from a surface.
}
\examples{
T_eval = seq(from=300, to=500, by=20);
k = c(1000, -20, 0.002);
pv = vapor_pressure(T_eval, k);
evaporation(pv, 10, T_eval, 150.2);
}
\author{
Peter Norgard
}
