% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluiddensity_TS.R
\name{fluiddensity_TS}
\alias{fluiddensity_TS}
\title{Seawater Density Calculator}
\usage{
fluiddensity_TS(tempC, sal, range = "low")
}
\arguments{
\item{tempC}{temperature in degrees Celsius; valid for -2 to 40 degrees C ("low" range from Millero & Poisson) or 0 to 180 degrees C ("high" range from Sharqawy et al)}

\item{sal}{practical salinity in g/kg; valid for 0 to 42 g/kg ("low" range from Millero & Poisson) or 0 to 160 g/kg ("high" range from Sharqawy et al)}

\item{range}{either "low" for Millero & Poisson or "high" for Sharqawy et al}
}
\value{
density in kg/m^3
}
\description{
Calculates fluid density as a function of temperature in Celsius and salinity in g/kg, using equations developed for seawater
}
\details{
This function contains two empirical equations, one from:
F. J. Millero, and A. Poisson, International one-atmosphere equation of state of seawater. Deep-Sea Research, 28A (6), 625 – 629, 1981.
and another from:
Sharqawy, M. H., Lienhard, J. H., & Zubair, S. M. (2010). Thermophysical properties of seawater: a review of existing correlations and data. Desalination and Water Treatment, 16(1–3), 354–380.
}
\examples{
rho_sw1 <- seawaterdensity(25,35)
rho_sw2 <- seawaterdensity(c(25,30,35),35)

}
