% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetic_pressure.R
\name{kinetic_pressure}
\alias{kinetic_pressure}
\title{Kinetic pressure
Compute the pressure resulting from a low-density gas in contact with a rigid three dimensional boundary.}
\usage{
kinetic_pressure(ng, Tg)
}
\arguments{
\item{ng, }{gas density in m^{-3}}

\item{Tg, }{average gas temperature in Kelvin}
}
\value{
Gas pressure arising from kinetic interaction of a low density gas obeying Boltzmann statistics in three dimensions.
}
\description{
Kinetic pressure
Compute the pressure resulting from a low-density gas in contact with a rigid three dimensional boundary.
}
\author{
Peter Norgard
}
