% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_rate_in.R
\name{particle_rate_in}
\alias{particle_rate_in}
\title{Particle rate in}
\usage{
particle_rate_in(ng, Ts, Tg, A_surface, mat)
}
\arguments{
\item{Ts}{numeric value; surface temperature \link{K}}

\item{Tg}{numeric vector; gas temperature \link{K}}

\item{A_surface}{numeric value; surface area of evaporating substance \link{m^2}}

\item{mat}{list; list of material properties for evaporating substance}
}
\description{
Compute the influx rate of particles due to material evaporation.
}
\author{
Peter Norgard
}
