% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_rate_out.R
\name{particle_rate_out}
\alias{particle_rate_out}
\title{Particle rate out}
\usage{
particle_rate_out(ng, Tg, A_orfice, mat)
}
\arguments{
\item{Tg}{numeric vector; gas temperature \link{K}}

\item{mat}{list; list of material properties for exiting substance}

\item{A_orifice}{numeric value; surface area of exit orifice \link{m^2}}
}
\description{
Compute the outflux rate of particles due to an aperture.
}
\author{
Peter Norgard
}
