% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seawaterdynamicviscosity.R
\name{seawaterdynamicviscosity}
\alias{seawaterdynamicviscosity}
\title{Seawater Dynamic Viscosity Calculator}
\usage{
seawaterdynamicviscosity(tempC, sal)
}
\arguments{
\item{tempC}{temperature in degrees Celsius; valid for 10 to 180 degrees C}

\item{sal}{practical salinity in g/kg; valid for 0 to 150 g/kg}
}
\value{
dynamic viscosity in Pa*s
}
\description{
Calculates seawater dynamic viscosity as a function of temperature in Celsius and salinity in g/kg
}
\details{
Function from J. D. Isdale, C. M. Spence, and J. S. Tudhope,
Physical properties of sea water solutions: viscosity,
Desalination, 10(4), 319 - 328, 1972.
}
\examples{
mu_sw <- seawaterdynamicviscosity(25,35)

}
