% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapor_pressure.R
\name{vapor_pressure}
\alias{vapor_pressure}
\title{Vapor pressure}
\usage{
vapor_pressure(Tval, k)
}
\arguments{
\item{k}{numeric vector; Antoine coefficients}

\item{T}{numeric vector; temperature(s) \link{K} at which to evaluate the vapor pressure(s)}
}
\description{
Compute the vapor pressure of a material using Antoine parameters. This function uses 3-valued or 4-valued Antoine parameters, and returns the vapor pressure in Pascals.
}
\examples{
T_eval = seq(from=300, to=500, by=20);
k = c(1000, -20, 0.002);
vapor_pressure(T_eval, k);
}
\author{
Peter Norgard
}
